requirements.txt
alembic==1.0.9
astroid==2.2.5
autopep8==1.4.4
blinker==1.4
certifi==2019.3.9
chardet==3.0.4
Click==7.0
dominate==2.3.5
Flask==1.0.2
Flask-Bootstrap==3.3.7.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.4.0
Flask-Script==2.0.6
Flask-SimpleMDE==0.3.0
Flask-SQLAlchemy==2.4.0
Flask-Uploads==0.2.1
Flask-WTF==0.14.2
gunicorn==19.9.0
idna==2.8
isort==4.3.17
itsdangerous==1.1.0
Jinja2==2.10.1
lazy-object-proxy==1.3.1
Mako==1.0.9
markdown2==2.3.7
MarkupSafe==1.1.1
mccabe==0.6.1
Pillow==6.0.0
psycopg2==2.8.2
pycodestyle==2.5.0
pylint==2.3.1
python-dateutil==2.8.0
python-editor==1.0.4
requests==2.21.0
six==1.12.0
SQLAlchemy==1.3.3
typed-ast==1.3.4
urllib3==1.24.2
visitor==0.1.3
Werkzeug==0.15.2
wrapt==1.11.1
WTForms==2.2.1
